backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.10.0
pydantic-settings==2.6.0
PyYAML==6.0.2

# AI/ML
openai==1.58.1
google-generativeai==0.8.0
langchain==0.3.14
langgraph==0.2.60
langchain-openai==0.2.14
langchain-google-genai==2.0.8
tiktoken==0.8.0

# Database - Updated for compatibility
supabase==2.7.4
httpx==0.27.0

# WebSocket
websockets==12.0
python-multipart==0.0.6

# Utilities
python-jose[cryptography]==3.3.0
